refactor(renderer/pages/game-details): colocated select-folder files to one feature folder
