+++ src/renderer/src/pages/game-details/select-folder/index.ts
@@ -1,0 +1,1 @@
+export * from "./select-folder-modal";
