+++ src/renderer/src/pages/game-details/select-folder/select-folder-modal.css.tsx
@@ -1,5 +1,5 @@
 import { style } from "@vanilla-extract/css";
-import { SPACING_UNIT, vars } from "../../theme.css";
+import { SPACING_UNIT, vars } from "../../../theme.css";
 
 export const container = style({
   display: "flex",
